fix: 내 식물 전체 조회 쿼리 수정
