+++ domain/core/src/main/kotlin/dnd11th/blooming/core/repository/image/ImageRepository.kt
@@ -12,26 +12,6 @@ import org.springframework.data.repository.query.Param
 interface ImageRepository : JpaRepository<Image, Long> {
     fun findAllByMyPlant(myPlant: MyPlant): List<Image>
 
-    @Query(
-        """
-    SELECT new dnd11th.blooming.core.entity.myplant.MyPlantWithImageUrl(mp, i.url)
-    FROM Image i
-    JOIN i.myPlant mp
-	JOIN FETCH mp.location l
-    WHERE mp.user = :user
-    AND i.favorite = true
-    AND i.updatedDate = (
-        SELECT MAX(i2.updatedDate)
-        FROM Image i2
-        WHERE i2.myPlant = mp
-        AND i2.favorite = true
-    )
-	""",
-    )
-    fun findMyPlantAndMostRecentFavoriteImageByUser(
-        @Param("user") user: User,
-    ): List<MyPlantWithImageUrl>
-
     @Modifying
     @Query(
         """
